requirements.txt
pypandoc==1.7.4
python-frontmatter==1.0.0
PyYAML==6.0
